+++ src/main/java/at/hannibal2/skyhanni/config/features/Garden.java
@@ -24,48 +24,79 @@ public class Garden {
     public Position skyMartCopperPricePos = new Position(44, -108, false, true);
 
     @Expose
-    @ConfigOption(name = "Visitor Helper", desc = "")
+    @ConfigOption(name = "Visitor", desc = "")
     @ConfigEditorAccordion(id = 1)
-    public boolean visitorHelper = false;
+    public boolean visitor = false;
 
     @Expose
-    @ConfigOption(name = "Visitor Display", desc = "Show all items needed for the visitors.")
+    @ConfigOption(name = "Visitor Timer", desc = "")
+    @ConfigAccordionId(id = 1)
+    @ConfigEditorAccordion(id = 2)
+    public boolean visitorTimer = false;
+
+    @Expose
+    @ConfigOption(name = "Visitor Timer", desc = "Timer when the next visitor will appear," +
+            "and a number how many visitors are already waiting.")
     @ConfigEditorBoolean
-    @ConfigAccordionId(id = 1)
-    public boolean visitorHelperDisplay = true;
+    @ConfigAccordionId(id = 2)
+    public boolean visitorTimerEnabled = true;
 
     @Expose
-    @ConfigOption(name = "Visitor Helper Position", desc = "")
-    @ConfigEditorButton(runnableId = "visitorHelper", buttonText = "Edit")
+    @ConfigOption(name = "Visitor Timer Position", desc = "")
+    @ConfigEditorButton(runnableId = "visitorTimer", buttonText = "Edit")
+    @ConfigAccordionId(id = 2)
+    public Position visitorTimerPos = new Position(0, 0, false, true);
+
+    @Expose
+    @ConfigOption(name = "Visitor Items Needed", desc = "")
     @ConfigAccordionId(id = 1)
-    public Position visitorHelperPos = new Position(0, 0, false, true);
+    @ConfigEditorAccordion(id = 3)
+    public boolean visitorNeeds = false;
+
+    @Expose
+    @ConfigOption(name = "Items Needed", desc = "Show all items needed for the visitors.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 3)
+    public boolean visitorNeedsDisplay = true;
+
+    @Expose
+    @ConfigOption(name = "Items Needed Position", desc = "")
+    @ConfigEditorButton(runnableId = "visitorNeeds", buttonText = "Edit")
+    @ConfigAccordionId(id = 3)
+    public Position visitorNeedsPos = new Position(0, 0, false, true);
+
+    @Expose
+    @ConfigOption(name = "Only when Close", desc = "Only show the needed items when close to the visitors.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 3)
+    public boolean visitorNeedsOnlyWhenClose = false;
 
     @Expose
     @ConfigOption(name = "Highlight Ready", desc = "Highlight the visitor when the required items are in the inventory.")
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 1)
-    public boolean visitorHelperHighlightReady = true;
+    public boolean visitorHighlightReady = true;
 
     @Expose
     @ConfigOption(name = "Show Price", desc = "Show the bazaar price of the items required for the visitors.")
     @ConfigEditorBoolean
     @ConfigAccordionId(id = 1)
-    public boolean visitorHelperShowPrice = true;
+    public boolean visitorShowPrice = true;
 
     @Expose
     @ConfigOption(name = "Numbers", desc = "")
-    @ConfigEditorAccordion(id = 2)
+    @ConfigEditorAccordion(id = 4)
     public boolean numbers = false;
 
     @Expose
-    @ConfigOption(name = "Crop Milestone", desc = "Show the number of the crop milestone in the inventory.")
+    @ConfigOption(name = "Crop Milestone", desc = "Show the number of crop milestones in the inventory.")
     @ConfigEditorBoolean
-    @ConfigAccordionId(id = 2)
-    public boolean cropMilestoneNumber = true;
+    @ConfigAccordionId(id = 4)
+    public boolean numberCropMilestone = true;
 
     @Expose
     @ConfigOption(name = "Crop Upgrades", desc = "Show the number of upgrades in the crop upgrades inventory.")
     @ConfigEditorBoolean
-    @ConfigAccordionId(id = 2)
-    public boolean cropUpgradesNumber = true;
+    @ConfigAccordionId(id = 4)
+    public boolean numberCropUpgrades = true;
 }