+++ src/main/java/at/hannibal2/skyhanni/events/TabListUpdateEvent.kt
@@ -1,0 +1,3 @@
+package at.hannibal2.skyhanni.events
+
+class TabListUpdateEvent(val tabList: List<String>): LorenzEvent()