+++ src/main/java/at/hannibal2/skyhanni/features/garden/GardenInventoryNumbers.kt
@@ -6,7 +6,7 @@ import at.hannibal2.skyhanni.events.RenderItemTipEvent
 import at.hannibal2.skyhanni.utils.InventoryUtils
 import at.hannibal2.skyhanni.utils.ItemUtils.getLore
 import at.hannibal2.skyhanni.utils.LorenzUtils
-import at.hannibal2.skyhanni.utils.NumberUtil.romanToDecimal
+import at.hannibal2.skyhanni.utils.NumberUtil.romanToDecimalIfNeeded
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import java.util.regex.Pattern
 
@@ -20,17 +20,18 @@ class GardenInventoryNumbers {
         if (!isEnabled()) return
 
         if (InventoryUtils.openInventoryName() == "Crop Milestones") {
-            if (!SkyHanniMod.feature.garden.cropMilestoneNumber) return
+            if (!SkyHanniMod.feature.garden.numberCropMilestone) return
 
             event.stack.getLore()
                 .map { patternTierProgress.matcher(it) }
                 .filter { it.matches() }
-                .map { it.group(1).romanToDecimal() - 1 }
+
+                .map { it.group(1).romanToDecimalIfNeeded() - 1 }
                 .forEach { event.stackTip = "" + it }
         }
 
         if (InventoryUtils.openInventoryName() == "Crop Upgrades") {
-            if (!SkyHanniMod.feature.garden.cropUpgradesNumber) return
+            if (!SkyHanniMod.feature.garden.numberCropUpgrades) return
 
             event.stack.getLore()
                 .map { patternUpgradeTier.matcher(it) }