+++ src/main/java/at/hannibal2/skyhanni/features/garden/GardenVisitorFeatures.kt
@@ -2,6 +2,7 @@ package at.hannibal2.skyhanni.features.garden
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.data.IslandType
+import at.hannibal2.skyhanni.events.GuiRenderEvent
 import at.hannibal2.skyhanni.events.InventoryOpenEvent
 import at.hannibal2.skyhanni.events.PacketEvent
 import at.hannibal2.skyhanni.events.withAlpha
@@ -16,16 +17,13 @@ import net.minecraft.client.Minecraft
 import net.minecraft.entity.EntityLivingBase
 import net.minecraft.network.play.client.C02PacketUseEntity
 import net.minecraft.network.play.server.S13PacketDestroyEntities
-import net.minecraftforge.client.event.RenderGameOverlayEvent
 import net.minecraftforge.event.entity.player.ItemTooltipEvent
 import net.minecraftforge.fml.common.eventhandler.EventPriority
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import net.minecraftforge.fml.common.gameevent.TickEvent
-import java.util.regex.Pattern
 
 class GardenVisitorFeatures {
 
-    private val pattern = Pattern.compile("(.*)§8x(.*)")
     private val visitors = mutableMapOf<String, Visitor>()
     private val display = mutableListOf<String>()
     private var lastClickedNpc = 0
@@ -34,8 +32,8 @@ class GardenVisitorFeatures {
     @SubscribeEvent
     fun onChatPacket(event: InventoryOpenEvent) {
         if (!isEnabled()) return
-        if (!SkyHanniMod.feature.garden.visitorHelperDisplay &&
-            !SkyHanniMod.feature.garden.visitorHelperHighlightReady
+        if (!SkyHanniMod.feature.garden.visitorNeedsDisplay &&
+            !SkyHanniMod.feature.garden.visitorHighlightReady
         ) return
 
         val npcItem = event.inventory.items[13] ?: return
@@ -57,11 +55,8 @@ class GardenVisitorFeatures {
             if (line == "§7Items Required:") continue
             if (line.isEmpty()) break
 
-            val matcher = pattern.matcher(line)
-            if (!matcher.matches()) continue
-
-            val itemName = matcher.group(1).trim()
-            val amount = matcher.group(2).toInt()
+            val (itemName, amount) = ItemUtils.readItemAmount(line)
+            if (itemName == null) continue
             visitor.items[itemName] = amount
         }
 
@@ -83,7 +78,7 @@ class GardenVisitorFeatures {
         }
         if (requiredItems.isEmpty()) return
 
-        display.add("Visitors need:")
+        display.add("Visitor Items Needed:")
         for ((name, amount) in requiredItems) {
             display.add(" -$name §8x$amount")
         }
@@ -95,7 +90,7 @@ class GardenVisitorFeatures {
     fun onTooltip(event: ItemTooltipEvent) {
         if (!isEnabled()) return
         if (!nearby) return
-        if (!SkyHanniMod.feature.garden.visitorHelperShowPrice) return
+        if (!SkyHanniMod.feature.garden.visitorShowPrice) return
 
         val name = event.itemStack.name ?: return
         if (name != "§aAccept Offer") return
@@ -109,17 +104,14 @@ class GardenVisitorFeatures {
             if (line == "") {
                 if (amountDifferentItems > 1) {
                     val format = NumberUtil.format(totalPrice)
-                    list[1] = list[1] + "$line §f(§6$format§f)"
+                    list[1] = list[1] + "$line §f(§6Total §6$format§f)"
                 }
                 break
             }
 
             if (i > 1) {
-                val matcher = pattern.matcher(line)
-                if (matcher.matches()) {
-                    val itemName = matcher.group(1).trim()
-                    val amount = matcher.group(2).toInt()
-
+                val (itemName, amount) = ItemUtils.readItemAmount(line)
+                if (itemName != null) {
                     val internalName = NEUItems.getInternalNameByName(itemName)
                     val auctionManager = NotEnoughUpdates.INSTANCE.manager.auctionManager
                     val lowestBin = auctionManager.getBazaarOrBin(internalName, false)
@@ -137,15 +129,31 @@ class GardenVisitorFeatures {
     @SubscribeEvent
     fun onTick(event: TickEvent.ClientTickEvent) {
         if (!isEnabled()) return
-        if (!SkyHanniMod.feature.garden.visitorHelperDisplay &&
-            !SkyHanniMod.feature.garden.visitorHelperHighlightReady &&
-            !SkyHanniMod.feature.garden.visitorHelperShowPrice
+        if (!SkyHanniMod.feature.garden.visitorNeedsDisplay &&
+            !SkyHanniMod.feature.garden.visitorHighlightReady &&
+            !SkyHanniMod.feature.garden.visitorShowPrice
         ) return
         if (tick++ % 60 != 0) return
 
-        nearby = LocationUtils.playerLocation().distance(LorenzVec(8.4, 72.0, -14.1)) < 10
+        val defaultVanillaSkin = LorenzVec(8.4, 72.0, -14.1)
+        val castleSkin = LorenzVec(-5, 75, 18)
+        val bambooSkin = LorenzVec(-12, 72, -25)
+        val hiveSkin = LorenzVec(-17, 71, -19)
+        val cubeSkin = LorenzVec(-17, 71, -19)
 
-        if (nearby && SkyHanniMod.feature.garden.visitorHelperHighlightReady) {
+        // TODO Only check current one, ignore others.
+        val list = mutableListOf<LorenzVec>()
+        list.add(defaultVanillaSkin)
+        list.add(castleSkin)
+        list.add(bambooSkin)
+        list.add(hiveSkin)
+        list.add(cubeSkin)
+
+        val playerLocation = LocationUtils.playerLocation()
+        nearby = list.map { playerLocation.distance(it) < 15 }.any { it }
+
+
+        if (nearby && SkyHanniMod.feature.garden.visitorHighlightReady) {
             checkVisitorsReady()
         }
     }
@@ -167,7 +175,7 @@ class GardenVisitorFeatures {
                 if (entity is EntityLivingBase) {
                     val color = LorenzColor.GREEN.toColor().withAlpha(120)
                     RenderLivingEntityHelper.setEntityColor(entity, color)
-                    { SkyHanniMod.feature.garden.visitorHelperHighlightReady }
+                    { SkyHanniMod.feature.garden.visitorHighlightReady }
                 }
             }
         }
@@ -188,26 +196,29 @@ class GardenVisitorFeatures {
         }
     }
 
+    // TODO make event
     @SubscribeEvent
     fun onSendEvent(event: PacketEvent.SendEvent) {
         val packet = event.packet
         if (packet !is C02PacketUseEntity) return
 
         val theWorld = Minecraft.getMinecraft().theWorld
-        val entity = packet.getEntityFromWorld(theWorld)
+        val entity = packet.getEntityFromWorld(theWorld) ?: return
         val entityId = entity.entityId
 
         lastClickedNpc = entityId
     }
 
     @SubscribeEvent
-    fun onRenderOverlay(event: RenderGameOverlayEvent.Post) {
-        if (event.type != RenderGameOverlayEvent.ElementType.ALL) return
+    fun onRenderOverlay(event: GuiRenderEvent) {
         if (!isEnabled()) return
-        if (!SkyHanniMod.feature.garden.visitorHelperDisplay) return
-        if (!nearby) return
+        if (!SkyHanniMod.feature.garden.visitorNeedsDisplay) return
 
-        SkyHanniMod.feature.garden.visitorHelperPos.renderStrings(display)
+        if (SkyHanniMod.feature.garden.visitorNeedsOnlyWhenClose) {
+            if (!nearby) return
+        }
+
+        SkyHanniMod.feature.garden.visitorNeedsPos.renderStrings(display)
     }
 
     class Visitor(val entityId: Int, val items: MutableMap<String, Int> = mutableMapOf())