+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 + Added **Crop Milestone** Number - Show the number of the crop milestone in the inventory.
 + Added **Crop Upgrades** Number - Show the number of upgrades in the crop upgrades inventory.
 + Added **Yaw Snapping** - Make it harder to overshoot certain yaw angles. (contrib nea89)
++ Added **Visitor Timer** - Timer when the next visitor will appear, and a number how many visitors are already waiting.
 
 ### Features from other Mods
 > *The following features are only there because I want them when testing SkyHanni features without other mods present.*